test: Extract testing models to a file
