+++ spec/prezzo/composable_spec.rb
@@ -1,35 +1,5 @@
 require "spec_helper"
 
-class FooCalculator
-  include Prezzo::Calculator
-
-  def formula
-    10.0
-  end
-end
-
-class BarCalculator
-  include Prezzo::Calculator
-
-  param :bar_param
-
-  def formula
-    bar_param
-  end
-end
-
-class ComposedCalculator
-  include Prezzo::Calculator
-
-  param :a_param
-  component :foo, FooCalculator
-  component :bar, BarCalculator
-
-  def formula
-    foo + bar
-  end
-end
-
 RSpec.describe Prezzo::Composable do
   let(:context) { { a_param: 42.3, bar_param: 15.3 } }
   let(:calculator) { ComposedCalculator.new(context) }
